add roberta and move bert_module (#788)
* add roberta and move bert_module


* add missing files
