+++ nemo/collections/nlp/modules/common/__init__.py
@@ -1,2 +1,3 @@
-from nemo.collections.nlp.modules.common.token_classifier import TokenClassifier
+from nemo.collections.nlp.modules.common.bert_module import *
 from nemo.collections.nlp.modules.common.huggingface import *
+from nemo.collections.nlp.modules.common.token_classifier import TokenClassifier
