+++ nemo/collections/nlp/modules/common/huggingface/__init__.py
@@ -1,0 +1,3 @@
+from nemo.collections.nlp.modules.common.huggingface.albert import *
+from nemo.collections.nlp.modules.common.huggingface.bert import *
+from nemo.collections.nlp.modules.common.huggingface.roberta import *
